feat: stats por campeonato no LineupBuilder + card redesign + locked_cost fix
- LineupBuilder: substitui colunas estáticas por stats/jogo do campeonato atual
  (PTS/G, K/G, DMG/G, ASS/G, WIN, SOBREV/G) via endpoint /stages/{id}/prior-stats
- Card de jogador: redesign 2 colunas (logo | info), texto responsivo com container
  queries (clamp), botão capitão apenas estrela sem multiplicador no card
- locked_cost: Integer → Numeric(6,2) (migration 0023) para preservar decimais
- PlayerStatsPage: fix PTS SOBREV com float artifacts (Math.round)
- Todas as colunas de stats centralizadas (LineupBuilder e PlayerStatsPage)
- Tooltips nos cabeçalhos das colunas de stats no LineupBuilder

+++ app/models/lineup.py
@@ -94,8 +94,8 @@ class LineupPlayer(Base):
         Boolean, nullable=False, server_default=text("false"),
         comment="True para o titular escolhido como capitão (multiplicador ×1.3)",
     )
-    locked_cost: Mapped[Optional[int]] = mapped_column(
-        Integer, nullable=True, comment="Cost at the moment of lock"
+    locked_cost: Mapped[Optional[Decimal]] = mapped_column(
+        Numeric(6, 2), nullable=True, comment="Cost at the moment of lock"
     )
     points_earned: Mapped[Optional[Decimal]] = mapped_column(
         Numeric(8, 2), nullable=True,
